completed #4 and cleaned up code :)

+++ tenmo-server/src/main/java/com/techelevator/tenmo/model/TransferDto.java
@@ -28,7 +28,7 @@ public class TransferDto {
         return amount;
     }
 
-    public void setAmount(int amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
 }
